split valid sequence output and crf into their own files
